model_zoo/vgae/requirements.txt
mindspore-gpu==1.7.0
numpy==1.21.6
mindspore-gl-gpu